Refactor project structure. - Use singular words. - Refactor directory structure. - Clean unused imports. - Remove sample files.
