+++ src/main/java/net/swmaestro/portal/auth/controller/AuthController.java
@@ -1,8 +1,6 @@
-package net.swmaestro.portal.auth;
+package net.swmaestro.portal.auth.controller;
 
 import net.swmaestro.portal.auth.service.AuthService;
-import net.swmaestro.portal.sample.HomeController;
-import net.swmaestro.portal.sample.service.SampleService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
@@ -11,9 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
