+++ src/main/java/net/swmaestro/portal/comment/CommentApi.java
@@ -17,7 +17,7 @@ import java.util.List;
 @javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-10-05T11:30:19.659Z")
 
 @Api(value = "comments", description = "the comments API")
-public interface CommentsApi {
+public interface CommentApi {
 
     @ApiOperation(value = "Delete Comment", notes = "Delete a comment.", response = Void.class, authorizations = {
         @Authorization(value = "smp_auth", scopes = {
