+++ src/main/java/net/swmaestro/portal/comment/CommentApiController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 @javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-10-05T11:30:19.659Z")
 
 @Controller
-public class CommentsApiController implements CommentsApi {
+public class CommentApiController implements CommentApi {
 
     public ResponseEntity<Void> deleteComment(
 @ApiParam(value = "Comment's ID",required=true ) @PathVariable("commentId") Integer commentId
