+++ src/main/java/net/swmaestro/portal/common/configuration/SwaggerConfig.java
@@ -1,9 +1,7 @@
-package net.swmaestro.portal.configuration;
+package net.swmaestro.portal.common.configuration;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
