+++ src/main/java/net/swmaestro/portal/lecture/controller/LectureApiController.java
@@ -2,29 +2,23 @@ package net.swmaestro.portal.lecture.controller;
 
 import io.swagger.annotations.ApiParam;
 import net.swmaestro.portal.auth.JWTAuthentication;
-import net.swmaestro.portal.auth.TokenUtil;
 import net.swmaestro.portal.lecture.service.LectureService;
 import net.swmaestro.portal.lecture.vo.Lecture;
 import net.swmaestro.portal.user.vo.User;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.token.Token;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 
 @javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-10-05T11:30:19.659Z")
 
 @Controller
-public class LecturesApiController implements LecturesApi {
+public class LectureApiController implements LectureApi {
 
     @Resource(name = "lectureService")
     private LectureService lectureService;
