+++ src/main/java/net/swmaestro/portal/user/controller/UserApiController.java
@@ -9,7 +9,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.annotation.Resource;
 import java.util.HashMap;
@@ -20,7 +19,7 @@ import java.util.Map;
 @javax.annotation.Generated(value = "class io.swagger.codegen.languages.SpringCodegen", date = "2016-10-05T11:30:19.659Z")
 
 @Controller
-public class UsersApiController implements UsersApi {
+public class UserApiController implements UserApi {
 
     @Resource(name="userService")
     private UserService userService;
