--- src/main/resources/mapper/sample/Sample_SQL.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="sample">
-	
-	<select id="selectUserNameList" parameterType="hashmap" resultType="hashmap">
-		<![CDATA[
-			SELECT
-			    user_name
-			FROM
-			    soma.soma_user
-			ORDER BY user_id DESC
-		]]>
-	</select>
-
-	<select id="selectUser" parameterType="hashmap" resultType="hashmap">
-		<![CDATA[
-			SELECT
-			    user_name
-			FROM
-			    soma.soma_user
-			ORDER BY user_id DESC
-		]]>
-	</select>
-	
-</mapper>
